Fix misreading of the CR30 community configuration changes
When `MIN_SOFTWARE_ENDSTOP_Y` is not defined, then the value for `Y_MIN_POS` is ignored. The Y axis can lower infinitely as far as the firmware is concerned.

Set `Y_MIN_POS` to 0 so that during the machine leveling, setting the new home position makes true home match Y == 0.